msm: Fix possible incorrect memcpy param in sensors ADSP driver
Using a newer gcc we get the following warning:

arch/arm/mach-msm/sensors_adsp.c: In function 'sns_ocmem_map_send':
arch/arm/mach-msm/sensors_adsp.c:877:39: warning: argument to 'sizeof' in
'memcpy' call is the same expression as the source; did you mean to
dereference it? [-Wsizeof-pointer-memaccess]

As vector is a pointer the sizeof(vector) is probably not what is desired,
change it to the sizeof the type of vector.

Change-Id: I8039901f2bea4ba6946d3934bac908816e41ee84
Signed-off-by: Kumar Gala <galak@codeaurora.org>