CHERRY-PICK: Revert "CHROMIUM: pci: default d3cold_allowed to false"
This reverts commit 96b63c80804860a1d70c2096d3cc300606a08f85.

Not using D3cold is causing a number of unexpected behaviors with the XHCI
driver.  I think it is better to fix the underlying issue that prompted
this to be disabled originally which was aborted suspends due to wakeups.
I believe that I have a firmware fix for that issue now.

BUG=chrome-os-partner:21663,chrome-os-partner:21524
TEST=manual: boot on falco with no USB devices attached and see
that the XHCI controller is put in D3cold state.

Change-Id: I79e2bc2093085645d68138f135855869badbb6b2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/65237
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66745
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Jay Kim <yongjaek@chromium.org>
Tested-by: Jay Kim <yongjaek@chromium.org>