CHROMIUMOS: mwifiex: Add debugfs file for testing reset of card
This is helpful in testing that the reset works.

Note that you need to be a little careful because there seems to be a
race where trying to reset the card while it's partway through the
reset process can eventually crash the system (memory corruption?).
Waiting 10 seconds between resets seems to work fairly reliably.

BUG=chrome-os-partner:21479
TEST=Run this for a while:
  while true; do
    sleep 10;
    echo 'y' > /sys/kernel/debug/mwifiex/mlan0/reset;
  done

Signed-off-by: Paul Stewart <pstew@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Originally-Reviewed-on: https://gerrit.chromium.org/gerrit/65265

[Forward port to kernel-3.10/wireless-3.8]

Change-Id: Idddf2487e038ffd9ae81c0039f88e36e030c72d9
Reviewed-on: https://chromium-review.googlesource.com/193988
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>