BACKPORT 3.10.15: xhci: Fix race between ep halt and URB cancellation
commit 526867c3ca0caa2e3e846cb993b0f961c33c2abb upstream.

The halted state of a endpoint cannot be cleared over CLEAR_HALT from a
user process, because the stopped_td variable was overwritten in the
handle_stopped_endpoint() function. So the xhci_endpoint_reset() function will
refuse the reset and communication with device can not run over this endpoint.
https://bugzilla.kernel.org/show_bug.cgi?id=60699

Signed-off-by: Florian Wolter <wolly84@web.de>
Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Cc: Jonghwan Choi <jhbird.choi@samsung.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

[benzh: Cherry-picked to usb-xhci topic branch because
cba17ddd504448452bd44c3071f6e0d91e521b03 depends on it]
Signed-off-by: Ben Zhang <benzh@chromium.org>

Change-Id: I1bdeda3c6807c927849079d53c63eade641240ef