net: qfec: Refactored driver to reduce Rx INTs
The problem is being addressed by this patch
by disabling the next Rx INT for 150 us when
a Rx INT occurs. The packet which caused the
INT is recieved & passed to the stack in the
ISR itself, hence there is no packet latency
for the packet which caused the INT. A hrtimer
will be scheduled to execute after 150 us in
the ISR when a Rx INT is recieved. The call-
back registered with the hrtimer will poll
for packets; if it finds at-least one packet
then the hrtimer is scheduled again to execute
after 150 us but if no packets are recieved
while executing the call-back then the Rx INT
is re-enabled.

Change-Id: I676cebe0f7bfce72f7d7592451b5c2208a52e6fa
Acked-by: John Nicholas <jnichola@qti.qualcomm.com>
Signed-off-by: Venkatesh Yadav Abbarapu <quicvenkat@codeaurora.org>