ARM: fix erratum 798181 workaround for multi-cluster systems
In a multi-cluster system, it is possible that the CPU doing a TLB
invalidation may not be affected by Cortex-A15 erratum 798181, but
other CPUs are.  Introduce the global has_erratum_a15_798181 which
is set if any CPUs in the system are affected by the erratum.  If
it is set, we'll broadcast the A15 erratum IPI regardless of the
CPU that's performing the TLB flush.

BUG=chrome-os-partner:21715
TEST=mem-stress-switcher

Original-Change-Id: I8b2ab8c6b6bffe28b265c1e11de8a11c42b75ee0
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/66533
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
(cherry picked from commit 00fd8cdaeef9a94b1824f2ef97e2209ec0ab2854)

Change-Id: I2e2d005f66d7d2d5013586bf04e68e9c15bde51d
Reviewed-on: https://gerrit.chromium.org/gerrit/66959
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>