gpio: avoid panic on NULL desc->chip in gpiod_request
Avoid calling desc_to_gpio if desc->chip is NULL, as this will
cause a kernel panic.

In the code above this, there is a test for chip==NULL, which
comes to the 'done' label if true.  In this case, the code
panics, since desc_to_gpio() uses desc->chip to look up the
gpio number.

The possibility of calling desc_to_gpio() with a NULL desc->chip
is present in other locations (gpiod_direction_input,
gpiod_direction_output, gpiod_set_debounce) which this fix
doesn't catch.

Cc: stable@vger.kernel.org
Signed-off-by: Tim Bird <tim.bird@sonymobile.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>