vfs: allow umount to handle mountpoints without revalidating them
Christopher reported a regression where he was unable to unmount a NFS
filesystem where the root had gone stale.  The problem is that
d_revalidate handles the root of the filesystem differently from other
dentries, but d_weak_revalidate does not.  We could simply fix this by
making d_weak_revalidate return success on IS_ROOT dentries, but there are
cases where we do want to revalidate the root of the fs.

A umount is really a special case.  We generally aren't interested in
anything but the dentry and vfsmount that's attached at that point.  If
the inode turns out to be stale we just don't care since the intent is to
stop using it anyway.

Try to handle this situation better by treating umount as a special case
in the lookup code.  Have it resolve the parent using normal means, and
then do a lookup of the final dentry without revalidating it.  In most
cases, the final lookup will come out of the dcache, but the case where
there's a trailing symlink or !LAST_NORM entry on the end complicates
things a bit.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Cc: Neil Brown <neilb@suse.de>
Reported-by: Christopher T Vogan <cvogan@us.ibm.com>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Cc: <stable@vger.kernel.org>	[3.9+]
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>