tile: Replace __get_cpu_var uses
__get_cpu_var() is used for multiple purposes in the kernel source. One
of them is address calculation via the form &__get_cpu_var(x). This
calculates the address for the instance of the percpu variable of the
current processor based on an offset.

Others usage cases are for storing and retrieving data from the current
processors percpu area.  __get_cpu_var() can be used as an lvalue when
writing data or on the right side of an assignment.

__get_cpu_var() is defined as :

__get_cpu_var() always only does a address determination. However, store
and retrieve operations could use a segment prefix (or global register
on other platforms) to avoid the address calculation.

this_cpu_write() and this_cpu_read() can directly take an offset into
a percpu area and use optimized assembly code to read and write per
cpu variables.

This patch converts __get_cpu_var into either an explicit address
calculation using this_cpu_ptr() or into a use of this_cpu operations
that use the offset. Thereby address calcualtions are avoided and less
registers are used when code is generated.

The patchset includes passes over all arches as well. Once these
operations are used throughout then specialized macros can be defined in
non -x86 arches as well in order to optimize per cpu access by f.e. using
a global register that may be set to the per cpu base.

Transformations done to __get_cpu_var()

1. Determine the address of the percpu instance of the current processor.

        DEFINE_PER_CPU(int, y);
        int *x = &__get_cpu_var(y);

    Converts to

        int *x = this_cpu_ptr(&y);

2. Same as #1 but this time an array structure is involved.

        DEFINE_PER_CPU(int, y[20]);
        int *x = __get_cpu_var(y);

    Converts to

        int *x = this_cpu_ptr(y);

3. Retrieve the content of the current processor's instance of a per
   cpu variable.

        DEFINE_PER_CPU(int, u);
        int x = __get_cpu_var(y)

   Converts to

        int x = __this_cpu_read(y);

4. Retrieve the content of a percpu struct

        DEFINE_PER_CPU(struct mystruct, y);
        struct mystruct x = __get_cpu_var(y);

   Converts to

        memcpy(this_cpu_ptr(&y), x, sizeof(x));

5. Assignment to a per cpu variable

        DEFINE_PER_CPU(int, y)
        __get_cpu_var(y) = x;

   Converts to

        this_cpu_write(y, x);

6. Increment/Decrement etc of a per cpu variable

        DEFINE_PER_CPU(int, y);
        __get_cpu_var(y)++

   Converts to

        this_cpu_inc(y)

Signed-off-by: Christoph Lameter <cl@linux.com>
Signed-off-by: Chris Metcalf <cmetcalf@tilera.com> [with minor cleanups]