tile: support KVM host mode
This commit enables the host side of KVM support for tilegx.

KVM support on tilegx presumes a client that runs at privilege level 1
(PL1), above normal user programs, which continue to run at PL0, but
below the normal (host) Linux, which runs at PL2.  Omitting all "trap
and emulate" support both simplifies the host as well as allowing a
paravirtualized guest to run with better overall performance; we may
in the future elect to add emulation support.

We don't support huge pages, or any of the tile-specific APIs that
require being locked to specific cores (e.g., UDN hardwall).

The Tilera booter/hypervisor (which runs at PL3) has been extended to
support a guest context (used for interrupts at PL1) as well as a
virtualization context that is applied to map guest PAs to real PAs.
Note that the eventual plan for the Tilera software stack is to
migrate away from having the Tilera booter provide PL3 functionality
like this and instead provide it all in the host kernel at PL2.
Meanwhile, the kvm host provides the Tilera hypervisor API to the
paravirtualized guest.

The commit adds a KVM_EXIT_xxx code, KVM_EXIT_AGAIN, which is used to
exit out to the host kernel, but not all the way out to qemu.  This is
helpful if we are trying to handle resched, sigpending, etc., but don't
need to end up back in userspace first.

Signed-off-by: Chris Metcalf <cmetcalf@tilera.com>