tile: enable building as a paravirtualized KVM_GUEST
This commit enables a special configure option to build the kernel to
run at PL1.  In this mode, the client can run under a KVM host kernel;
it can also run under the older Tilera hypervisor that ran the
operating system at PL1 by default.

The PL1 kernel runs with half the virtual address space and half the
physical address space of the PL2 kernel, so <asm/page.h> reflects
modifying those constants appropriately.  We make some things a little
more generic ("intrpt1" references from the old PL1 kernel
nomenclature have now been normalized to "intrpt"), and simplify some
other nomenclature (using MEM_SV_START instead of MEM_SV_INTRPT to
reflect the fact that where the supervisor starts is the fact of
interest, not that it happens to start with the interrupt vectors).

The simulator support for reflecting Elf binary data back out
associated with simulator backtrace information needs some additional
extension.  It currently will report backtrace information for the
guest kernel, but not for processes running within the guest kernel;
additional work in the simulator is required to be able to provide the
necessary path information for that to work.  For now we disable
simulator notifications within the guest kernel.

The timer interrupt for the guest uses the AUX_TILE_TIMER hardware,
leaving the regular TILE_TIMER for the host.

Signed-off-by: Chris Metcalf <cmetcalf@tilera.com>