CHROMIUM: Handle EC LPC driver removal properly.
We were registering the device on module init, but weren't removing it on
exit. Now we do.

BUG=none
BRANCH=none
TEST=manual

Build the driver as a module (CONFIG_MFD_CROS_EC_LPC=m), then

  find /sys -exec ls -ldF {} \+ > /tmp/before
  modprobe cros_ec_lpc

  rmmod cros_ec_lpc
  find /sys -exec ls -ldF {} \+ > /tmp/after

  diff /tmp/before /tmp/after

There might be quite a lot of differences but none of them should be
related to the cros_ec_lpc driver. For example, you shouldn't see

  /sys/bus/platfor/devices/cros-ec-*
  /sys/bus/platfor/drivers/cros_ec_lpc/*

Change-Id: Iccec7ea04d909530577043b926f9d96d887b1de9
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172638
Reviewed-by: Benson Leung <bleung@chromium.org>