UPSTREAM: coredump: add new %P variable in core_pattern
Add a new %P variable to be used in core_pattern.  This variable contains
the global PID (PID in the init namespace) as %p contains the PID in the
current namespace which isn't always what we want.

The main use for this is to make it easier to handle crashes that happened
within a container.  With that new variables it's possible to have the
crashes dumped into the container or forwarded to the host with the right
PID (from the host's point of view).

BUG=chromium:351568
TEST=`cbuildbot link-release` passes

Signed-off-by: Stphane Graber <stgraber@ubuntu.com>
Reported-by: Hans Feldt <hans.feldt@ericsson.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Eric W. Biederman <ebiederm@xmission.com>
Cc: Andy Whitcroft <apw@canonical.com>
Acked-by: Serge E. Hallyn <serge.hallyn@ubuntu.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Mike Frysinger <vapier@chromium.org>
(cherry picked from commit 65aafb1e7484b7434a0c1d4c593191ebe5776a2f)

Change-Id: I3521e781c226c9fb73b16be6f207fd0dcece9229
Previous-Reviewed-on: https://chromium-review.googlesource.com/190020
(cherry picked from commit d93d643d22e470fa2438d778e29773c0c30ae436)
Reviewed-on: https://chromium-review.googlesource.com/190533
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>