clk: tegra: move audio clk to common file
Move audio clocks and PLLA initialization to a common file so it can be used by
multiple Tegra SoCs. Also a new array tegra114_clks is introduced for Tegra114
which specifies which common clocks are available on Tegra114 and what their
DT IDs are.

Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>

http://patchwork.ozlabs.org/patch/285206/
BUG=None
TEST=Verified booting to console

Change-Id: Ib350d734fadfc79dbc65def1f5249fe5950be106
[rklein: Removed CLK_SET_RATE_NO_REPARENT flag, as it isn't supported here]
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/173277
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>