msm: vidc: decode time info correction for erroneous case
In case of erroneous clip, profile information for erroneous
frames are not proper, this is because decode time for these
frames gets accumulated to next non erroneous frame, it causes
shoot-up in decode time information. This is fixed by calling
ddl_calc_core_proc_time function in case of decoder error.

Change-Id: I7cf08fa12cfbd8d874d93edc53860e4466422adb
Signed-off-by: Pushkaraj Patil <ppatil@codeaurora.org>