CHROMIUM: config: increased ramoops record size to 256K
RAMOOPS record size is currently set to 128K. Bump it up to 256K so that
more of the kernel log buffer can be preserved in case of panics and
reboots.

BUG=chromium:295127
TEST=do things that generate lots of kernel logs such as suspend/resume
with /sys/power/pm_print_times set to 1. Monitor dmesg to see if the
buffer has rotated around and reboot the system when it does. Do "ls -l
/dev/pstore/console-ramoops" and check that its size is > 128K.

Change-Id: I6ecbe18c055cd4d0b548aa3dd1cd4d3b96484aee
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/170128
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>